💡refactor: App.ts type 분리
