+++ App.tsx
@@ -4,12 +4,7 @@ import {createStackNavigator} from '@react-navigation/stack';
 import GitLoginScreen from './src/screens/GitLoginScreen';
 import Home from './src/screens/Home';
 import GitWebViewScreen from './src/screens/GitWebViewScreen';
-
-export type RootStackPramList = {
-  GitLoginScreen: {authCode: string} | undefined;
-  GitWebViewScreen: undefined;
-  Home: undefined;
-};
+import {RootStackPramList} from './type';
 
 const Stack = createStackNavigator<RootStackPramList>();
