+++ src/hooks/type.ts
@@ -1,5 +1,5 @@
 import {StackNavigationProp} from '@react-navigation/stack';
-import {RootStackPramList} from '../../App';
+import {RootStackPramList} from '../../type';
 
 export type useGitHubRedirectProps = {
   navigation: StackNavigationProp<RootStackPramList>;
