+++ type.ts
@@ -1,0 +1,5 @@
+export type RootStackPramList = {
+  GitLoginScreen: {authCode: string} | undefined;
+  GitWebViewScreen: undefined;
+  Home: undefined;
+};
